move routes/channel -> routes/channelSchema
